Refactor insert outbound shipment line struct/method
