+++ server/graphql/invoice_line/src/mutations/mod.rs
@@ -1,2 +1,3 @@
+pub mod common_insert;
 pub mod inbound_shipment_line;
 pub mod outbound_shipment_line;
