+++ server/service/src/invoice/outbound_shipment/batch.rs
@@ -2,11 +2,11 @@ use repository::{Invoice, InvoiceLine, RepositoryError};
 
 use crate::{
     invoice_line::{
+        common_insert_line::{InsertInvoiceLine, InsertInvoiceLineError},
         outbound_shipment_line::{
             delete_outbound_shipment_line, insert_outbound_shipment_line,
             update_outbound_shipment_line, DeleteOutboundShipmentLine,
-            DeleteOutboundShipmentLineError, InsertOutboundShipmentLine,
-            InsertOutboundShipmentLineError, UpdateOutboundShipmentLine,
+            DeleteOutboundShipmentLineError, UpdateOutboundShipmentLine,
             UpdateOutboundShipmentLineError,
         },
         outbound_shipment_service_line::{
@@ -37,7 +37,7 @@ use super::{
 #[derive(Clone, Debug)]
 pub struct BatchOutboundShipment {
     pub insert_shipment: Option<Vec<InsertOutboundShipment>>,
-    pub insert_line: Option<Vec<InsertOutboundShipmentLine>>,
+    pub insert_line: Option<Vec<InsertInvoiceLine>>,
     pub update_line: Option<Vec<UpdateOutboundShipmentLine>>,
     pub delete_line: Option<Vec<DeleteOutboundShipmentLine>>,
     pub insert_service_line: Option<Vec<InsertOutboundShipmentServiceLine>>,
@@ -54,12 +54,8 @@ pub struct BatchOutboundShipment {
 
 pub type InsertShipmentsResult =
     Vec<InputWithResult<InsertOutboundShipment, Result<Invoice, InsertOutboundShipmentError>>>;
-pub type InsertLinesResult = Vec<
-    InputWithResult<
-        InsertOutboundShipmentLine,
-        Result<InvoiceLine, InsertOutboundShipmentLineError>,
-    >,
->;
+pub type InsertLinesResult =
+    Vec<InputWithResult<InsertInvoiceLine, Result<InvoiceLine, InsertInvoiceLineError>>>;
 pub type UpdateLinesResult = Vec<
     InputWithResult<
         UpdateOutboundShipmentLine,
@@ -287,7 +283,7 @@ mod test {
         invoice::outbound_shipment::{
             BatchOutboundShipment, DeleteOutboundShipmentError, InsertOutboundShipment,
         },
-        invoice_line::outbound_shipment_line::InsertOutboundShipmentLine,
+        invoice_line::common_insert_line::InsertInvoiceLine,
         service_provider::ServiceProvider,
         InputWithResult,
     };
@@ -310,15 +306,13 @@ mod test {
                 input.id = "new_id".to_string();
                 input.other_party_id = mock_name_store_b().id;
             })]),
-            insert_line: Some(vec![inline_init(
-                |input: &mut InsertOutboundShipmentLine| {
-                    input.invoice_id = "new_id".to_string();
-                    input.id = "new_line_id".to_string();
-                    input.item_id = mock_item_a().id;
-                    input.stock_line_id = mock_stock_line_a().id;
-                    input.number_of_packs = 1.0
-                },
-            )]),
+            insert_line: Some(vec![inline_init(|input: &mut InsertInvoiceLine| {
+                input.invoice_id = "new_id".to_string();
+                input.id = "new_line_id".to_string();
+                input.item_id = mock_item_a().id;
+                input.stock_line_id = mock_stock_line_a().id;
+                input.number_of_packs = 1.0
+            })]),
             update_line: None,
             delete_line: None,
             update_shipment: None,
