+++ server/service/src/invoice_line/outbound_shipment_line/insert/generate.rs
@@ -1,90 +1,92 @@
-use crate::invoice::common::calculate_total_after_tax;
-use repository::{
-    InvoiceLineRow, InvoiceLineRowType, InvoiceRow, InvoiceRowStatus, ItemRow, StockLineRow,
-};
+// use crate::{
+//     invoice::common::calculate_total_after_tax,
+//     invoice_line::insert_invoice_lines_common::{InsertInvoiceLine, InsertInvoiceLineError},
+// };
+// use repository::{
+//     InvoiceLineRow, InvoiceLineRowType, InvoiceRow, InvoiceRowStatus, ItemRow, StockLineRow,
+// };
 
-use super::{InsertOutboundShipmentLine, InsertOutboundShipmentLineError};
+// pub fn generate(
+//     input: InsertInvoiceLine,
+//     item_row: ItemRow,
+//     batch: StockLineRow,
+//     invoice: InvoiceRow,
+// ) -> Result<(InvoiceLineRow, StockLineRow), InsertInvoiceLineError> {
+//     let adjust_total_number_of_packs = invoice.status == InvoiceRowStatus::Picked;
 
-pub fn generate(
-    input: InsertOutboundShipmentLine,
-    item_row: ItemRow,
-    batch: StockLineRow,
-    invoice: InvoiceRow,
-) -> Result<(InvoiceLineRow, StockLineRow), InsertOutboundShipmentLineError> {
-    let adjust_total_number_of_packs = invoice.status == InvoiceRowStatus::Picked;
+//     let update_batch = generate_batch_update(&input, batch.clone(), adjust_total_number_of_packs);
+//     let new_line = generate_line(input, item_row, batch, invoice);
 
-    let update_batch = generate_batch_update(&input, batch.clone(), adjust_total_number_of_packs);
-    let new_line = generate_line(input, item_row, batch, invoice);
+//     Ok((new_line, update_batch))
+// }
 
-    Ok((new_line, update_batch))
-}
+// fn generate_batch_update(
+//     input: &InsertInvoiceLine,
+//     batch: StockLineRow,
+//     adjust_total_number_of_packs: bool,
+// ) -> StockLineRow {
+//     let mut update_batch = batch;
 
-fn generate_batch_update(
-    input: &InsertOutboundShipmentLine,
-    batch: StockLineRow,
-    adjust_total_number_of_packs: bool,
-) -> StockLineRow {
-    let mut update_batch = batch;
+//     let reduction = input.number_of_packs;
 
-    let reduction = input.number_of_packs;
+//     update_batch.available_number_of_packs = update_batch.available_number_of_packs - reduction;
+//     if adjust_total_number_of_packs {
+//         update_batch.total_number_of_packs = update_batch.total_number_of_packs - reduction;
+//     }
 
-    update_batch.available_number_of_packs = update_batch.available_number_of_packs - reduction;
-    if adjust_total_number_of_packs {
-        update_batch.total_number_of_packs = update_batch.total_number_of_packs - reduction;
-    }
+//     update_batch
+// }
 
-    update_batch
-}
+// fn generate_line(
+//     InsertInvoiceLine {
+//         id,
+//         invoice_id,
+//         item_id,
+//         stock_line_id,
+//         number_of_packs,
+//         total_before_tax,
+//         tax: _,
+//         note: _,
+//     }: InsertInvoiceLine,
+//     ItemRow {
+//         name: item_name,
+//         code: item_code,
+//         ..
+//     }: ItemRow,
+//     StockLineRow {
+//         sell_price_per_pack,
+//         cost_price_per_pack,
+//         pack_size,
+//         batch,
+//         expiry_date,
+//         location_id,
+//         note,
+//         ..
+//     }: StockLineRow,
+//     InvoiceRow { tax, .. }: InvoiceRow,
+// ) -> InvoiceLineRow {
+//     let total_before_tax = total_before_tax.unwrap_or(cost_price_per_pack * number_of_packs as f64);
+//     let total_after_tax = calculate_total_after_tax(total_before_tax, tax);
 
-fn generate_line(
-    InsertOutboundShipmentLine {
-        id,
-        invoice_id,
-        item_id,
-        stock_line_id,
-        number_of_packs,
-        total_before_tax,
-        tax: _,
-    }: InsertOutboundShipmentLine,
-    ItemRow {
-        name: item_name,
-        code: item_code,
-        ..
-    }: ItemRow,
-    StockLineRow {
-        sell_price_per_pack,
-        cost_price_per_pack,
-        pack_size,
-        batch,
-        expiry_date,
-        location_id,
-        note,
-        ..
-    }: StockLineRow,
-    InvoiceRow { tax, .. }: InvoiceRow,
-) -> InvoiceLineRow {
-    let total_before_tax = total_before_tax.unwrap_or(cost_price_per_pack * number_of_packs as f64);
-    let total_after_tax = calculate_total_after_tax(total_before_tax, tax);
-
-    InvoiceLineRow {
-        id,
-        invoice_id,
-        item_id,
-        location_id,
-        pack_size,
-        batch,
-        expiry_date,
-        sell_price_per_pack,
-        cost_price_per_pack,
-        r#type: InvoiceLineRowType::StockOut,
-        number_of_packs,
-        item_name,
-        item_code,
-        stock_line_id: Some(stock_line_id),
-        total_before_tax,
-        total_after_tax,
-        tax,
-        note,
-        inventory_adjustment_reason_id: None,
-    }
-}
+//     InvoiceLineRow {
+//         id,
+//         invoice_id,
+//         item_id,
+//         location_id,
+//         pack_size,
+//         batch,
+//         expiry_date,
+//         sell_price_per_pack,
+//         cost_price_per_pack,
+//         r#type: InvoiceLineRowType::StockOut,
+//         number_of_packs,
+//         item_name,
+//         item_code,
+//         stock_line_id: Some(stock_line_id),
+//         total_before_tax,
+//         total_after_tax,
+//         tax,
+//         note,
+//         inventory_adjustment_reason_id: None,
+//     }
+// }
