+++ server/service/src/invoice_line/outbound_shipment_line/insert/mod.rs
@@ -1,32 +1,27 @@
-use crate::{invoice_line::query::get_invoice_line, service_provider::ServiceContext, WithDBError};
-use repository::{InvoiceLine, InvoiceLineRowRepository, RepositoryError, StockLineRowRepository};
+use crate::{
+    invoice_line::{
+        common_insert_line::{generate, validate, InsertInvoiceLine, InsertInvoiceLineError},
+        query::get_invoice_line,
+    },
+    service_provider::ServiceContext,
+};
+use repository::{InvoiceLine, InvoiceLineRowRepository, InvoiceRowType, StockLineRowRepository};
 
-mod generate;
-mod validate;
-
-use generate::generate;
-use validate::validate;
-#[derive(Clone, Debug, PartialEq, Default)]
-pub struct InsertOutboundShipmentLine {
-    pub id: String,
-    pub invoice_id: String,
-    pub item_id: String,
-    pub stock_line_id: String,
-    pub number_of_packs: f64,
-    pub total_before_tax: Option<f64>,
-    pub tax: Option<f64>,
-}
-
-type OutError = InsertOutboundShipmentLineError;
+type OutError = InsertInvoiceLineError;
 
 pub fn insert_outbound_shipment_line(
     ctx: &ServiceContext,
-    input: InsertOutboundShipmentLine,
+    input: InsertInvoiceLine,
 ) -> Result<InvoiceLine, OutError> {
     let new_line = ctx
         .connection
         .transaction_sync(|connection| {
-            let (item, invoice, batch) = validate(&input, &ctx.store_id, &connection)?;
+            let (item, invoice, batch) = validate(
+                &input,
+                &ctx.store_id,
+                &connection,
+                InvoiceRowType::OutboundShipment,
+            )?;
             let (new_line, update_batch) = generate(input, item, batch, invoice)?;
             InvoiceLineRowRepository::new(&connection).upsert_one(&new_line)?;
             StockLineRowRepository::new(&connection).upsert_one(&update_batch)?;
@@ -36,44 +31,6 @@ pub fn insert_outbound_shipment_line(
         })
         .map_err(|error| error.to_inner_error())?;
     Ok(new_line)
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub enum InsertOutboundShipmentLineError {
-    LineAlreadyExists,
-    DatabaseError(RepositoryError),
-    InvoiceDoesNotExist,
-    NotAnOutboundShipment,
-    NotThisStoreInvoice,
-    CannotEditFinalised,
-    ItemNotFound,
-    StockLineNotFound,
-    NumberOfPacksBelowOne,
-    LocationIsOnHold,
-    LocationNotFound,
-    StockLineAlreadyExistsInInvoice(String),
-    ItemDoesNotMatchStockLine,
-    NewlyCreatedLineDoesNotExist,
-    BatchIsOnHold,
-    ReductionBelowZero { stock_line_id: String },
-}
-
-impl From<RepositoryError> for InsertOutboundShipmentLineError {
-    fn from(error: RepositoryError) -> Self {
-        InsertOutboundShipmentLineError::DatabaseError(error)
-    }
-}
-
-impl<ERR> From<WithDBError<ERR>> for InsertOutboundShipmentLineError
-where
-    ERR: Into<InsertOutboundShipmentLineError>,
-{
-    fn from(result: WithDBError<ERR>) -> Self {
-        match result {
-            WithDBError::DatabaseError(error) => error.into(),
-            WithDBError::Error(error) => error.into(),
-        }
-    }
 }
 
 #[cfg(test)]
@@ -92,8 +49,9 @@ mod test {
 
     use crate::{
         invoice::outbound_shipment::{UpdateOutboundShipment, UpdateOutboundShipmentStatus},
-        invoice_line::outbound_shipment_line::{
-            insert::InsertOutboundShipmentLine, InsertOutboundShipmentLineError as ServiceError,
+        invoice_line::{
+            common_insert_line::InsertInvoiceLine,
+            common_insert_line::InsertInvoiceLineError as ServiceError,
         },
         service_provider::ServiceProvider,
     };
@@ -116,7 +74,7 @@ mod test {
         assert_eq!(
             service.insert_outbound_shipment_line(
                 &context,
-                inline_init(|r: &mut InsertOutboundShipmentLine| {
+                inline_init(|r: &mut InsertInvoiceLine| {
                     r.id = mock_outbound_shipment_a_invoice_lines()[0].id.clone();
                     r.invoice_id = mock_outbound_shipment_a_invoice_lines()[0]
                         .invoice_id
@@ -130,7 +88,7 @@ mod test {
         assert_eq!(
             service.insert_outbound_shipment_line(
                 &context,
-                inline_init(|r: &mut InsertOutboundShipmentLine| {
+                inline_init(|r: &mut InsertInvoiceLine| {
                     r.id = "new outbound shipment line id".to_string();
                     r.invoice_id = mock_outbound_shipment_a_invoice_lines()[0]
                         .invoice_id
@@ -146,7 +104,7 @@ mod test {
         assert_eq!(
             service.insert_outbound_shipment_line(
                 &context,
-                inline_init(|r: &mut InsertOutboundShipmentLine| {
+                inline_init(|r: &mut InsertInvoiceLine| {
                     r.id = "new outbound shipment line id".to_string();
                     r.invoice_id = "new invoice id".to_string();
                     r.item_id = mock_item_b_lines()[0].item_id.clone();
@@ -161,7 +119,7 @@ mod test {
         assert_eq!(
             service.insert_outbound_shipment_line(
                 &context,
-                inline_init(|r: &mut InsertOutboundShipmentLine| {
+                inline_init(|r: &mut InsertInvoiceLine| {
                     r.id = "new outbound line id".to_string();
                     r.invoice_id = "invalid".to_string();
                     r.number_of_packs = 1.0;
@@ -174,7 +132,7 @@ mod test {
         assert_eq!(
             service.insert_outbound_shipment_line(
                 &context,
-                inline_init(|r: &mut InsertOutboundShipmentLine| {
+                inline_init(|r: &mut InsertInvoiceLine| {
                     r.id = "new outbound line id".to_string();
                     r.invoice_id = mock_outbound_shipment_a_invoice_lines()[0]
                         .invoice_id
@@ -189,7 +147,7 @@ mod test {
         assert_eq!(
             service.insert_outbound_shipment_line(
                 &context,
-                inline_init(|r: &mut InsertOutboundShipmentLine| {
+                inline_init(|r: &mut InsertInvoiceLine| {
                     r.id = "new outbound line id".to_string();
                     r.invoice_id = mock_outbound_shipment_a_invoice_lines()[0]
                         .invoice_id
@@ -206,7 +164,7 @@ mod test {
         assert_eq!(
             service.insert_outbound_shipment_line(
                 &context,
-                inline_init(|r: &mut InsertOutboundShipmentLine| {
+                inline_init(|r: &mut InsertInvoiceLine| {
                     r.id = "new outbound line id".to_string();
                     r.invoice_id = mock_outbound_shipment_a_invoice_lines()[0]
                         .invoice_id
@@ -223,7 +181,7 @@ mod test {
         assert_eq!(
             service.insert_outbound_shipment_line(
                 &context,
-                inline_init(|r: &mut InsertOutboundShipmentLine| {
+                inline_init(|r: &mut InsertInvoiceLine| {
                     r.id = "new outbound line id".to_string();
                     r.invoice_id = mock_outbound_shipment_a_invoice_lines()[0]
                         .invoice_id
@@ -240,7 +198,7 @@ mod test {
         assert_eq!(
             service.insert_outbound_shipment_line(
                 &context,
-                inline_init(|r: &mut InsertOutboundShipmentLine| {
+                inline_init(|r: &mut InsertInvoiceLine| {
                     r.id = "new outbound line id".to_string();
                     r.invoice_id = mock_outbound_shipment_a_invoice_lines()[0]
                         .invoice_id
@@ -259,7 +217,7 @@ mod test {
         assert_eq!(
             service.insert_outbound_shipment_line(
                 &context,
-                inline_init(|r: &mut InsertOutboundShipmentLine| {
+                inline_init(|r: &mut InsertInvoiceLine| {
                     r.id = "new outbound line id".to_string();
                     r.invoice_id = mock_outbound_shipment_a_invoice_lines()[0]
                         .invoice_id
@@ -279,7 +237,7 @@ mod test {
         assert_eq!(
             service.insert_outbound_shipment_line(
                 &context,
-                inline_init(|r: &mut InsertOutboundShipmentLine| {
+                inline_init(|r: &mut InsertInvoiceLine| {
                     r.id = "new outbound line id".to_string();
                     r.invoice_id = mock_outbound_shipment_a_invoice_lines()[0]
                         .invoice_id
@@ -327,7 +285,7 @@ mod test {
         service
             .insert_outbound_shipment_line(
                 &context,
-                inline_init(|r: &mut InsertOutboundShipmentLine| {
+                inline_init(|r: &mut InsertInvoiceLine| {
                     r.id = "new outbound line id".to_string();
                     r.invoice_id = mock_outbound_shipment_c_invoice_lines()[0]
                         .invoice_id
@@ -378,7 +336,7 @@ mod test {
         service
             .insert_outbound_shipment_line(
                 &context,
-                inline_init(|r: &mut InsertOutboundShipmentLine| {
+                inline_init(|r: &mut InsertInvoiceLine| {
                     r.id = "new allocated invoice line".to_string();
                     r.invoice_id = mock_outbound_shipment_c_invoice_lines()[0]
                         .invoice_id
@@ -418,7 +376,7 @@ mod test {
         service
             .insert_outbound_shipment_line(
                 &context,
-                inline_init(|r: &mut InsertOutboundShipmentLine| {
+                inline_init(|r: &mut InsertInvoiceLine| {
                     r.id = "new picked invoice line".to_string();
                     r.invoice_id = mock_outbound_shipment_c_invoice_lines()[0]
                         .invoice_id
