+++ server/service/src/invoice_line/outbound_shipment_unallocated_line/allocate/generate.rs
@@ -11,7 +11,8 @@ use util::{
 };
 
 use crate::invoice_line::{
-    outbound_shipment_line::{InsertOutboundShipmentLine, UpdateOutboundShipmentLine},
+    common_insert_line::InsertInvoiceLine,
+    outbound_shipment_line::UpdateOutboundShipmentLine,
     outbound_shipment_unallocated_line::{
         DeleteOutboundShipmentUnallocatedLine, UpdateOutboundShipmentUnallocatedLine,
     },
@@ -20,7 +21,7 @@ use crate::invoice_line::{
 #[derive(Default)]
 pub struct GenerateOutput {
     pub update_lines: Vec<UpdateOutboundShipmentLine>,
-    pub insert_lines: Vec<InsertOutboundShipmentLine>,
+    pub insert_lines: Vec<InsertInvoiceLine>,
     pub update_unallocated_line: Option<UpdateOutboundShipmentUnallocatedLine>,
     pub delete_unallocated_line: Option<DeleteOutboundShipmentUnallocatedLine>,
     pub skipped_expired_stock_lines: Vec<StockLine>,
@@ -148,9 +149,9 @@ fn generate_new_line(
     invoice_id: &str,
     packs_to_allocate: f64,
     stock_line: &StockLine,
-) -> InsertOutboundShipmentLine {
+) -> InsertInvoiceLine {
     let stock_line_row = &stock_line.stock_line_row;
-    InsertOutboundShipmentLine {
+    InsertInvoiceLine {
         id: uuid::uuid(),
         invoice_id: invoice_id.to_string(),
         item_id: stock_line_row.item_id.clone(),
@@ -158,6 +159,7 @@ fn generate_new_line(
         number_of_packs: packs_to_allocate,
         total_before_tax: None,
         tax: None,
+        note: None,
     }
 }
